requirements.txt

fastapi==0.115.0
uvicorn==0.32.0
sqlmodel==0.0.22
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.9.0
requests==2.32.3
